Clean up Block Proof protobuf definitions * Fixed protobuf unused import warnings in several files * Removed the chain-of-trust-proof file from block stream (which does not use it) * Added the chain-of-trust-proof to the internal history protos (where it is used).     * Fixed all imports * Moved merkle sibling hashes from block proof to state proof     * Fixed imports to match * Extracted record file proof to its own proto file     * Fixed imports to match * Extracted TSS signature proof to its own proto file     * Fixed imports to match
