+++ hapi/hedera-protobuf-java-api/src/main/proto/block/stream/record_file_item.proto
@@ -20,7 +20,6 @@ option java_package = "com.hedera.hapi.block.stream.protoc";
 // <<<pbj.java_package = "com.hedera.hapi.block.stream">>> This comment is special code for setting PBJ Compiler java package
 option java_multiple_files = true;
 
-import "services/basic_types.proto";
 import "services/timestamp.proto";
 import "streams/sidecar_file.proto";
 import "streams/record_stream_file.proto";
