+++ hapi/hedera-protobuf-java-api/src/main/proto/block/stream/state_proof.proto
@@ -27,6 +27,7 @@ option java_multiple_files = true;
 
 import "services/timestamp.proto";
 import "block/stream/record_file_item.proto";
+import "block/stream/tss_signed_block_proof.proto";
 import "services/state/blockstream/merkle_leaf.proto";
 
 message StateProof {
@@ -50,32 +51,6 @@ message StateProof {
          */
         TssSignedBlockProof signed_block_proof = 2;
     }
-}
-
-/**
- * A proof containing a TSS signature for a single block.
- */
-message TssSignedBlockProof {
-    /**
-     * A TSS signature for one block.<br/>
-     * This is a single signature representing the collection of partial
-     * signatures from nodes holding strictly greater than 2/3 of the
-     * current network "weight" in aggregate. The signature is produced by
-     * cryptographic "aggregation" of the partial signatures to produce a
-     * single signature that can be verified with the network public key,
-     * but could not be produced by fewer nodes than required to meet the
-     * threshold for network stake "weight".
-     * <p>
-     * This message MUST make use of a threshold signature scheme like `BLS`
-     * which provides the necessary cryptographic guarantees.<br/>
-     * This signature SHALL use a TSS signature to provide a single signature
-     * that represents the consensus signature of consensus nodes.<br/>
-     * The exact subset of nodes that signed SHALL neither be known nor
-     * tracked, but it SHALL be cryptographically verifiable that the
-     * threshold was met if the signature itself can be validated with
-     * the network public key (a.k.a `LedgerID`).
-     */
-    bytes block_signature = 1;
 }
 
 /**
@@ -135,4 +110,37 @@ message SiblingNode {
      * computed hashes as the proof is validated.
      */
     bytes hash = 2;
+}
+
+
+/**
+ * A hash of a "sibling" to an entry in a Merkle tree.
+ *
+ * When constructing a binary merkle tree, each internal node is a hash
+ * constructed from the hash of two "descendant" nodes. Those two nodes
+ * are "siblings" and the order (first, second) in which the two hash values
+ * are combined affects the parent hash.<br/>
+ * This may be used to reconstruct a portion of a merkle tree starting from
+ * a node of interest up to the root of the tree.
+ */
+message MerkleSiblingHash {
+    /**
+     * A flag for the position of this sibling.
+     * <p>
+     * If this is set then this sibling MUST be the first hash in the pair of
+     * sibling hashes of a binary merkle tree.<br/>
+     * If this is unset, then this sibling MUST be the second hash in the pair
+     * of sibling hashes of a binary merkle tree.
+     */
+    bool is_first = 1;
+
+    /**
+     * A byte array of a sibling hash.<br/>
+     * This is the hash for the sibling at this point in the merkle tree.
+     * <p>
+     * The algorithm for this hash SHALL match the algorithm for the block that
+     * contains this sibling.<br/>
+     * This SHALL contain the raw (e.g.) 384 bits (48 bytes) of the hash value.
+     */
+    bytes sibling_hash = 2;
 }
