+++ hapi/hedera-protobuf-java-api/src/main/proto/services/auxiliary/hints/hints_key_publication.proto
@@ -13,8 +13,6 @@ syntax = "proto3";
 package com.hedera.hapi.services.auxiliary.hints;
 
 // SPDX-License-Identifier: Apache-2.0
-import "services/state/hints/hints_types.proto";
-
 option java_package = "com.hedera.hapi.services.auxiliary.hints.legacy";
 // <<<pbj.java_package = "com.hedera.hapi.services.auxiliary.hints">>> This comment is special code for setting PBJ Compiler java package
