+++ hapi/hedera-protobuf-java-api/src/main/proto/services/state/history/history_types.proto
@@ -4,7 +4,6 @@ package com.hedera.hapi.node.state.history;
 
 // SPDX-License-Identifier: Apache-2.0
 import "services/timestamp.proto";
-import "block/stream/chain_of_trust_proof.proto";
 
 option java_package = "com.hederahashgraph.api.proto.java";
 // <<<pbj.java_package = "com.hedera.hapi.node.state.history">>> This comment is special code for setting PBJ Compiler java package
@@ -86,7 +85,7 @@ message HistoryProof {
    * history's metadata. May be switched from AggregatedNodeSignatures
    * to a recursive proof when one becomes available.
    */
-  com.hedera.hapi.block.stream.ChainOfTrustProof chain_of_trust_proof = 3;
+  ChainOfTrustProof chain_of_trust_proof = 3;
   /**
    * If set, the uncompressed proof of chain of trust from the ledger id to
    * the target address book; the uncompressed version of the WRAPS proof is
@@ -283,3 +282,41 @@ message WrapsMessageHistory {
   repeated WrapsMessageDetails messages = 1;
 }
 
+message ChainOfTrustProof {
+    oneof proof {
+        /**
+         * If there is not yet a SNARK proving the chain of trust from ledger id to
+         * the hinTS verification key, an aggregation of Schnorr signatures on
+         * the concatenation of the ledger id and genesis hinTS verification key
+         * that serve as witnesses for the SNARK prover algorithm.
+         */
+        AggregatedNodeSignatures aggregated_node_signatures = 1;
+        /**
+         * If known, a ZK-compressed SNARK proof proving the chain of trust from
+         * the ledger id to this hinTS verification key.
+         */
+        bytes wraps_proof = 2;
+    }
+}
+
+/**
+ * An aggregation of node signatures on some data.
+ * <p>
+ * Can be used to prove the genesis hinTS verification key in a block proof; but
+ * not succinct and not recursive; hence in normal operations with TSS, used only
+ * until the first recursive proof is available.
+ */
+message AggregatedNodeSignatures {
+  /**
+   * The aggregated signature.
+   */
+  bytes aggregated_signature = 1;
+  /**
+   * In ascending order, the ids of the nodes that contributed signatures.
+   */
+  repeated uint64 signing_node_ids = 2;
+  /**
+   * The hinTS key that this signature witnesses.
+   */
+  bytes verification_key = 3;
+}
