+++ hapi/hedera-protobuf-java-api/src/main/proto/streams/sidecar_file.proto
@@ -26,7 +26,6 @@ option java_package = "com.hedera.services.stream.proto";
 // <<<pbj.java_package = "com.hedera.hapi.streams">>> This comment is special code for setting PBJ Compiler java package
 option java_multiple_files = true;
 
-import "services/basic_types.proto";
 import "services/timestamp.proto";
 import "streams/contract_state_change.proto";
 import "streams/contract_action.proto";
